Element functions are now split into multiple file and started the ground work for onPress support
