+++ front/views/settings/SettingsProfileView.tsx
@@ -90,11 +90,11 @@ const ProfileSettings = ({ navigation }: { navigation: any }) => {
 						{
 							type: "text",
 							title: "Email",
-							onPress: () => {
-								console.log("Go to email settings");
-							},
 							data: {
 								text: user.email || "Aucun email associé",
+								onPress: () => {
+									console.log("Go to email settings");
+								},
 							},
 						},
 						{
